Add an explanation to all Danger comments
People who're new to working on our repos might not know that we use
Danger, so we should be sure to explain what it is and why we use it.

We only want to display this message, though, when there are entries in
the `status_report` (which always contains a hash with keys `errors`,
`warnings`, `messages`, and `markdowns`, hence the flattening).

Note the trailing `\` on paragraph lines: these are so we don't end up
with paragraph breaks littered through the resulting paragraphs. This
is a problem because the `markdown` command uses [Kramdown's GitHub
Flavored Markdown parser](https://kramdown.gettalong.org/parser/gfm.html),
which defaults to maintaining hard line wraps. Ideally, we'd be able to
pass the `hard_wrap` option to `markdown` to fix this, but we can't for
now, so manually-collapsing is our only option.