Pin the danger gem to a known good major version
Open-ended dependency versions ('>= 0') prevent the gem from
building. Using '~> 0' caused dependency problems in projects
that try to install this gem. Also, for some reason, installing
this gem would use danger version 3, which is quite old.

This change requires danger version 5, which is the current
release, and has been tested by building this gem, and then
installing it both on a mature ruby project and on a test project
which does nothing but try to install this gem.